requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.33.0
datasets>=2.14.0
tokenizers>=0.13.3
accelerate>=0.21.0

# Parameter-Efficient Fine-Tuning
peft>=0.4.0
bitsandbytes>=0.41.0

# Evaluation Metrics
sacrebleu>=2.3.1
evaluate>=0.4.0
nltk>=3.8.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# API and Web Interface
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
gradio>=3.40.0
pydantic>=2.0.0

# Configuration and Logging
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0
hydra-core>=1.3.2

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
pre-commit>=3.3.3

# Visualization and Monitoring
matplotlib>=3.7.0
seaborn>=0.12.0
wandb>=0.15.0
tensorboard>=2.14.0

# Jupyter and Notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0
notebook>=7.0.0

# GPU Memory Optimization
flash-attn>=2.0.0; platform_system != "Darwin"

# Cloud Storage and APIs (optional)
google-cloud-storage>=2.10.0
boto3>=1.28.0

# Additional Utilities
tqdm>=4.65.0
requests>=2.31.0
python-multipart>=0.0.6

# Language Detection (optional)
langdetect>=1.0.9
polyglot>=16.7.4

# Medical/Healthcare NLP (optional for domain-specific features)
scispacy>=0.5.3
# en-core-sci-sm @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_core_sci_sm-0.5.3.tar.gz